requirements.txt
Django==4.2
djangorestframework==3.14.0
mysqlclient==2.1.1
django-mysql==4.9.0
django-cors-headers==3.14.0
pika==1.3.1
requests==2.28.2
Pillow==9.5.0